+++ client/src/components/layouts/dashboardLayout/DashboardLayoutRoute.js
@@ -1,18 +1,26 @@
 import React, { Fragment } from 'react';
-import Sidebar from './Sidebar';
+import Sidebar from '../layoutComponents/Sidebar';
+import Header from '../layoutComponents/Header';
 import { Container, Col, Row } from 'react-bootstrap';
 import { Route } from 'react-router-dom';
-import EditorSidebarData from '../editor/EditorSidebarData';
+import SidebarData from '../SidebarData';
 
+const headerData = {
+    title: "Researchify",
+    dashboardURL: "/dashboard",
+    profileURL: "/dashboard/profile"
+}
 
-const EditorLayout = ({ children }) => {
+const DashboardLayout = ({ children }) => {
+
 
     return (
         <Fragment>
+            <Header data={headerData} />
             <Container fluid>
                 <Row>
                     <Col className="sidebar-wrapper" md={2} lg={2} xl={1}>
-                        <Sidebar data={EditorSidebarData} />
+                        <Sidebar data={SidebarData} />
                     </Col>
                     <Col className="page-content-wrapper" md={10} lg={10} xl={11}>
                         {children}
@@ -20,23 +28,22 @@ const EditorLayout = ({ children }) => {
                 </Row>
             </Container>
         </Fragment>
-    )
-
+    );
 };
 
-const EditorLayoutRoute = ({ component: Component, ...rest }) => {
+const DashboardLayoutRoute = ({ component: Component, ...rest}) => {
     //get 'component' and renamed to 'Component', any other 'props' renamed to 'rest'
     return (
-        <Route
-            {...rest}
+        <Route 
+            {...rest} 
             render={routeProps => (
-                <EditorLayout>
+                <DashboardLayout>
                     <Component {...routeProps} />
-                </EditorLayout>
+                </DashboardLayout>
             )}
         />
     );
 };
 
 
-export default EditorLayoutRoute;
+export default DashboardLayoutRoute;
