+++ client/src/components/layouts/editorLayout/EditorLayoutRoute.js
@@ -1,27 +1,18 @@
 import React, { Fragment } from 'react';
-import Sidebar from './Sidebar';
-import Header from './Header';
+import Sidebar from './layoutComponents/Sidebar';
 import { Container, Col, Row } from 'react-bootstrap';
 import { Route } from 'react-router-dom';
-import SidebarData from './SidebarData';
-import EditorSideBarData from '../editor/EditorSideBarData'
+import EditorSidebarData from '../editor/EditorSidebarData';
 
-const headerData = {
-    title: "Researchify",
-    dashboardURL: "/dashboard",
-    profileURL: "/dashboard/profile"
-}
 
-const DashboardLayout = ({ children }) => {
-
+const EditorLayout = ({ children }) => {
 
     return (
         <Fragment>
-            <Header data={headerData} />
             <Container fluid>
                 <Row>
                     <Col className="sidebar-wrapper" md={2} lg={2} xl={1}>
-                        <Sidebar data={SidebarData} />
+                        <Sidebar data={EditorSidebarData} />
                     </Col>
                     <Col className="page-content-wrapper" md={10} lg={10} xl={11}>
                         {children}
@@ -29,22 +20,23 @@ const DashboardLayout = ({ children }) => {
                 </Row>
             </Container>
         </Fragment>
-    );
+    )
+
 };
 
-const DashboardLayoutRoute = ({ component: Component, ...rest}) => {
+const EditorLayoutRoute = ({ component: Component, ...rest }) => {
     //get 'component' and renamed to 'Component', any other 'props' renamed to 'rest'
     return (
-        <Route 
-            {...rest} 
+        <Route
+            {...rest}
             render={routeProps => (
-                <DashboardLayout>
+                <EditorLayout>
                     <Component {...routeProps} />
-                </DashboardLayout>
+                </EditorLayout>
             )}
         />
     );
 };
 
 
-export default DashboardLayoutRoute;
+export default EditorLayoutRoute;
